[quantization] Rename attn to attention
This commit renames attn to attention for consistency.

TICO-DCO-1.0-Signed-off-by: seongwoo <mhs4670go@naver.com>
