+++ tico/quantization/wrapq/wrappers/registry.py
@@ -49,11 +49,11 @@ _CORE_MODULES = (
     "tico.quantization.wrapq.wrappers.fairseq.quant_encoder_layer",
     "tico.quantization.wrapq.wrappers.fairseq.quant_mha",
     ## qwen_vl ##
-    "tico.quantization.wrapq.wrappers.qwen_vl.quant_text_attn",
+    "tico.quantization.wrapq.wrappers.qwen_vl.quant_text_attention",
     "tico.quantization.wrapq.wrappers.qwen_vl.quant_text_mlp",
     "tico.quantization.wrapq.wrappers.qwen_vl.quant_text_decoder_layer",
     "tico.quantization.wrapq.wrappers.qwen_vl.quant_text_model",
-    "tico.quantization.wrapq.wrappers.qwen_vl.quant_vision_attn",
+    "tico.quantization.wrapq.wrappers.qwen_vl.quant_vision_attention",
     "tico.quantization.wrapq.wrappers.qwen_vl.quant_vision_mlp",
     "tico.quantization.wrapq.wrappers.qwen_vl.quant_vision_patch_embed",
     "tico.quantization.wrapq.wrappers.qwen_vl.quant_vision_patch_merger",
